flask_backend/requirements.txt
flask==3.1.*
gunicorn==23.0.*
bcrypt==4.2.*
PyJWT==2.9.*
flask-cors==5.0.*
flask-limiter>=3.5
python-dotenv==1.1.*
psycopg2-binary==2.9.*
scikit-learn>=1.4
pandas>=2.1
numpy>=1.26
joblib>=1.3
Pillow>=10.0
pytesseract>=0.3.10
